+++ src/index.js
@@ -1,0 +1,20 @@
+import TmHelloWorld from "./components/tmHelloWorld"
+import TmField from "./components/tmField"
+
+const TendermintComponents = {
+  TmHelloWorld,
+  TmField
+}
+
+TendermintComponents.install = Vue =>
+  Object.keys(TendermintComponents)
+    .filter(component => component !== "install")
+    .map(component => TendermintComponents[component])
+    .forEach(Vue.use)
+
+export default TendermintComponents
+
+export {
+  default as TmHelloWorld
+} from "./components/tmHelloWorld/tmHelloWorld.vue"
+export { default as TmField } from "./components/tmField/tmField.vue"