+++ tests/unit/tmHelloWorld.spec.js
@@ -1,5 +1,5 @@
 import { shallowMount } from "@vue/test-utils"
-import HelloWorld from "@/components/HelloWorld"
+import HelloWorld from "@/components/tmHelloWorld/tmHelloWorld.vue"
 
 describe("HelloWorld.vue", () => {
   it("renders props.msg when passed", () => {